Added setup for input handling shader (commented out)
Made some minor changes to existing code and added a bunch of new stuff to prepare for a new shader that applies actions.

It isn't functional yet because I couldn't get the creation of a bind group for it to work, I couldn't find any examples.
Also I discovered the `AsBindGroup` trait, which I should've used from the start.
So that will be applied first...